+++ src/cmd/browse/views/deck.rs
